scripts/requirements.txt
python-dotenv>=1.0.0
anthropic>=0.39.0
openai>=1.50.0
zhipuai>=2.1.0
